notion-to-google-calendar-sync/requirements.txt
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz